Restructure entire project according to latest Gradle
