--- build.gradle
@@ -1,28 +1,0 @@
-buildscript {
-    ext {
-        kotlin_version = '1.3.50'
-
-        junit_version = '4.12'
-        mockito_version = '2.28.2'
-    }
-
-    repositories {
-        mavenCentral()
-    }
-
-    dependencies {
-        classpath 'org.jetbrains.dokka:dokka-gradle-plugin:0.9.18'
-    }
-}
-
-subprojects {
-    buildscript {
-        repositories {
-            mavenCentral()
-        }
-    }
-
-    repositories {
-        mavenCentral()
-    }
-}
