+++ core/src/main/kotlin/net/gouline/kapsule/Kapsules.kt
