+++ core/src/test/kotlin/net/gouline/kapsule/TestModules.kt
