+++ gradle.properties
@@ -1,3 +1,6 @@
+org.gradle.parallel=true
+org.gradle.caching=true
+
 # Project
 publish.groupId=net.gouline.kapsule
 publish.version=0.1-SNAPSHOT
