+++ gradle/libs.versions.toml
@@ -1,0 +1,10 @@
+[versions]
+junit-jupiter-engine = "5.10.3"
+mockito-kotlin = "5.4.0"
+
+[libraries]
+junit-jupiter-engine = { module = "org.junit.jupiter:junit-jupiter-engine", version.ref = "junit-jupiter-engine" }
+mockito-kotlin = { module = "org.mockito.kotlin:mockito-kotlin", version.ref = "mockito-kotlin" }
+
+[plugins]
+kotlin-jvm = { id = "org.jetbrains.kotlin.jvm", version = "2.0.0" }
