--- kapsule-core/build.gradle
@@ -1,21 +1,0 @@
-buildscript {
-    ext.kotlin_version = kotlin_version
-
-    dependencies {
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-    }
-}
-
-apply plugin: 'kotlin'
-
-group project['publish.groupId']
-version project['publish.version']
-
-dependencies {
-    compile "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
-
-    testCompile "junit:junit:$junit_version"
-    testCompile "org.mockito:mockito-core:$mockito_version"
-}
-
-apply from: '../gradle/publish.gradle'
