+++ sample/README.md
@@ -1,0 +1,6 @@
+# Sample
+
+Sample CLI application to demonstrate the concept of context and how it can be applied to dependency injection.
+
+Contains unit tests with split modules for test dependencies.
+
