+++ sample/build.gradle.kts
@@ -1,0 +1,30 @@
+plugins {
+    alias(libs.plugins.kotlin.jvm)
+    application
+}
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    implementation(project(":core"))
+
+    testImplementation(libs.junit.jupiter.engine)
+    testImplementation(libs.mockito.kotlin)
+    testImplementation("org.jetbrains.kotlin:kotlin-test-junit5")
+}
+
+application {
+    mainClass = "net.gouline.kapsule.demo.DemoKt"
+}
+
+java {
+    toolchain {
+        languageVersion = JavaLanguageVersion.of(17)
+    }
+}
+
+tasks.named<Test>("test") {
+    useJUnitPlatform()
+}
