+++ sample/src/main/kotlin/net/gouline/kapsule/demo/Context.kt
