+++ sample/src/main/kotlin/net/gouline/kapsule/demo/Demo.kt
