+++ sample/src/main/kotlin/net/gouline/kapsule/demo/mock/Dependencies.kt
