+++ sample/src/test/kotlin/net/gouline/kapsule/demo/TestContext.kt
