+++ sample/src/test/kotlin/net/gouline/kapsule/demo/di/TestModules.kt
