--- samples/README.md
@@ -1,16 +1,0 @@
-# Kapsule Samples
-
-This directory contains sample projects showcasing some of Kapsule use cases.
-
-## android
-
-Android application based on the MVP pattern with one screen and mixed modules providing plain and Android-specific objects to showcase how Kapsule can be used in a simple app. 
-
-Contains Robolectric unit tests with split modules for test dependencies.
-
-## simple
-
-Trivial command-line application that prints out some lines of text and demonstrates the concept of context and how it can be applied to dependency injection. 
-
-Contains unit tests with split modules for test dependencies.
-
