--- samples/simple/build.gradle
@@ -1,23 +1,0 @@
-buildscript {
-    ext.kotlin_version = kotlin_version
-
-    dependencies {
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-    }
-}
-
-plugins {
-    id 'application'
-}
-
-apply plugin: 'kotlin'
-
-mainClassName = 'net.gouline.kapsule.demo.DemoKt'
-
-dependencies {
-    compile project(':kapsule-core')
-
-    compile "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
-
-    testCompile "junit:junit:$junit_version"
-}
