--- samples/simple/src/test/kotlin/net/gouline/kapsule/demo/DemoAppTestCase.kt
@@ -1,31 +1,0 @@
-/*
- * Copyright 2017 Mike Gouline
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
-
-package net.gouline.kapsule.demo
-
-import junit.framework.TestCase
-import org.junit.Test
-
-/**
- * Test case for [Demo].
- */
-class DemoAppTestCase : TestCase() {
-
-    private val testContext = TestContext()
-
-    @Test fun testInjection() {
-        val demo = Demo(testContext)
-        demo.auth.authenticate()
-        assertNull(demo.userDao.firstName)
-        assertNull(demo.userDao.lastName)
-        assertNull(demo.userDao.email)
-        assertEquals(0, demo.stateDao.code)
-    }
-}
