--- settings.gradle
@@ -1,10 +1,0 @@
-rootProject.name = 'kapsule'
-
-include 'kapsule-core'
-
-if (System.getenv('EXCLUDE_SAMPLES')) {
-    println('Excluding sample projects')
-} else {
-    println('Including sample projects')
-    include 'samples:simple', 'samples:android'
-}
