+++ settings.gradle.kts
@@ -1,0 +1,6 @@
+plugins {
+    id("org.gradle.toolchains.foojay-resolver-convention") version "0.8.0"
+}
+
+rootProject.name = "kapsule"
+include("core", "sample")
